fix(initImageSet): upload image to cache and remove images from cache
State 0: The cache is uninitialized and the directory associated does not exist
State 1: The user has taken a single image, the cache directory and URI array has exactly one element
State 2: The user taken or deleted multiple images, the user can delete images any where in the sequence but new images are always
inserted at the front of the array.
State 3: The user has deleted all the images taken and once the user pressed remove and the cache directory and URI array both had exactly one element remaining,
the directory was deleted and the array was set to empty
State 4: The user pressed out of the modal and if the directory existed it was deleted and the URI array was set to empty.